Bug 1304: Adjusting the stamp driver and translator units to ensure a proper value passing
git-svn-id: file:///tmp/svnroot_only_rec/internal/rec-project/rec/trunk@5877 8c4c7310-a39f-45c4-bd0d-b10b2e5d99a4

+++ experiments/condensadorcilindrico/src/java/server/pt/utl/ist/elab/driver/condensadorcilindrico/CCStampDriver.java
@@ -64,16 +64,16 @@ public class CCStampDriver extends AbstractStampDriver {
 
         stampConfig.addCommandData(
                 StampConfigTranslator.CALIBRATE_STR,
-                new Float(Defaults.defaultIfEmpty(calibrateConfig, calibrateInfo)));
+                new Integer(Defaults.defaultIfEmpty(calibrateConfig, calibrateInfo)));
 
         stampConfig.addCommandData(
                 StampConfigTranslator.START_POS_STR,
-                new Float(Defaults.defaultIfEmpty(initPosConfig, initPosInfo)));
+                new Integer(Defaults.defaultIfEmpty(initPosConfig, initPosInfo)));
         
         
         stampConfig.addCommandData(
                 StampConfigTranslator.END_POS_STR,
-                new Float(Defaults.defaultIfEmpty(endPosConfig, endPosInfo)));
+                new Integer(Defaults.defaultIfEmpty(endPosConfig, endPosInfo)));
 
         stampConfig.addCommandData(StampConfigTranslator.NUMSAMPLES_STR, new Integer(totalSamples));
